Rangoli
Rangoli, one of the most beautiful and most pleasing art forms of India, is comprised of two words, 'rang' meaning 'color' and 'aavalli' meaning colored creepers' or 'row of colors'. Rangoli basically comprises of the art of making designs or patterns on the walls or the floor of the house, using finely ground white powder along with different colors. Numerous households in the Indian subcontinent make use of Rangoli designs for decorating the courtyard of their house.